Add sign out function and redirect to personal profile page after log in
